requirements.txt
flask
pandas
openpyxl
python-barcode
pillow
gunicorn